feat(service): export `hyper::service::MakeServiceRef`
It's sealed, and has a blanket implementation, and so should only be
used as bounds. Even still, its hidden from the docs.
