+++ src/service/mod.rs
@@ -33,7 +33,8 @@ mod make_service;
 mod new_service;
 mod service;
 
-pub use self::make_service::{make_service_fn, MakeService};
+pub use self::make_service::{make_service_fn, MakeService, MakeServiceRef};
+// NewService is soft-deprecated.
 #[doc(hidden)]
 pub use self::new_service::NewService;
 pub use self::service::{service_fn, service_fn_ok, Service};
